requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
colorama==0.4.6
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
datasets==2.14.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
evaluate==0.4.1
executing==2.0.0
fastjsonschema==2.19.0
filelock==3.12.4
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.16.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.5
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
safetensors==0.4.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
terminado==0.18.0
tinycss2==1.2.1
tokenizers==0.14.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.35.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
widgetsnbextension==4.0.9
xxhash==3.4.1
yarl==1.9.2
